chore: separate folder structure
