+++ packages/wallets/src/features/cashier/components/flows/index.ts
@@ -1,0 +1,3 @@
+export * from './WalletDeposit';
+export * from './WalletTransactions';
+export * from './WalletTransfer';
