+++ packages/wallets/src/features/cashier/components/screens/WalletDepositScreens/WalletDepositCryptoCurrencyDetails/WalletDepositCryptoCurrencyDetails.scss
