+++ packages/wallets/src/features/cashier/components/screens/WalletDepositScreens/WalletDepositFiat/WalletDepositFiat.scss
