+++ packages/wallets/src/features/cashier/components/screens/WalletDepositScreens/index.ts
@@ -1,0 +1,5 @@
+export * from './WalletDepositCrypto';
+export * from './WalletDepositCryptoAddress';
+export * from './WalletDepositCryptoCurrencyDetails';
+export * from './WalletDepositCryptoDisclaimers';
+export * from './WalletDepositFiat';
