+++ packages/wallets/src/features/cashier/components/screens/WalletTransferScreens/WalletTransferFormAccountCard/WalletTransferFormAccountCard.scss
